+++ psutil/_psutil_windows.c
@@ -229,14 +229,12 @@ psutil_pid_exists(PyObject *self, PyObject *args)
     long pid;
     int status;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
 
     status = psutil_pid_is_running(pid);
-    if (-1 == status) {
+    if (-1 == status)
         return NULL; // exception raised in psutil_pid_is_running()
-    }
     return PyBool_FromLong(status);
 }
 
@@ -253,13 +251,11 @@ psutil_pids(PyObject *self, PyObject *args)
     PyObject *pid = NULL;
     PyObject *retlist = PyList_New(0);
 
-    if (retlist == NULL) {
+    if (retlist == NULL)
         return NULL;
-    }
     proclist = psutil_get_pids(&numberOfReturnedPIDs);
-    if (NULL == proclist) {
+    if (proclist == NULL)
         goto error;
-    }
 
     for (i = 0; i < numberOfReturnedPIDs; i++) {
         pid = Py_BuildValue("I", proclist[i]);
@@ -292,12 +288,10 @@ psutil_proc_kill(PyObject *self, PyObject *args)
     HANDLE hProcess;
     long pid;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
-    if (pid == 0) {
+    if (pid == 0)
         return AccessDenied();
-    }
 
     hProcess = OpenProcess(PROCESS_TERMINATE, FALSE, pid);
     if (hProcess == NULL) {
@@ -335,12 +329,10 @@ psutil_proc_wait(PyObject *self, PyObject *args)
     long pid;
     long timeout;
 
-    if (! PyArg_ParseTuple(args, "ll", &pid, &timeout)) {
+    if (! PyArg_ParseTuple(args, "ll", &pid, &timeout))
         return NULL;
-    }
-    if (pid == 0) {
+    if (pid == 0)
         return AccessDenied();
-    }
 
     hProcess = OpenProcess(SYNCHRONIZE | PROCESS_QUERY_INFORMATION,
                            FALSE, pid);
@@ -395,15 +387,12 @@ psutil_proc_cpu_times(PyObject *self, PyObject *args)
     HANDLE      hProcess;
     FILETIME    ftCreate, ftExit, ftKernel, ftUser;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
 
     hProcess = psutil_handle_from_pid(pid);
-    if (hProcess == NULL) {
+    if (hProcess == NULL)
         return NULL;
-    }
-
     if (! GetProcessTimes(hProcess, &ftCreate, &ftExit, &ftKernel, &ftUser)) {
         CloseHandle(hProcess);
         if (GetLastError() == ERROR_ACCESS_DENIED) {
@@ -439,32 +428,6 @@ psutil_proc_cpu_times(PyObject *self, PyObject *args)
 
 
 /*
- * Alternative implementation of the one above but bypasses ACCESS DENIED.
- */
-static PyObject *
-psutil_proc_cpu_times_2(PyObject *self, PyObject *args)
-{
-    DWORD pid;
-    PSYSTEM_PROCESS_INFORMATION process;
-    PVOID buffer;
-    double user, kernel;
-
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
-        return NULL;
-    }
-    if (! psutil_get_proc_info(pid, &process, &buffer)) {
-        return NULL;
-    }
-    user = (double)process->UserTime.HighPart * 429.4967296 + \
-           (double)process->UserTime.LowPart * 1e-7;
-    kernel = (double)process->KernelTime.HighPart * 429.4967296 + \
-             (double)process->KernelTime.LowPart * 1e-7;
-    free(buffer);
-    return Py_BuildValue("(dd)", user, kernel);
-}
-
-
-/*
  * Return a Python float indicating the process create time expressed in
  * seconds since the epoch.
  */
@@ -478,20 +441,16 @@ psutil_proc_create_time(PyObject *self, PyObject *args)
     BOOL        ret;
     FILETIME    ftCreate, ftExit, ftKernel, ftUser;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
 
     // special case for PIDs 0 and 4, return system boot time
-    if (0 == pid || 4 == pid) {
+    if (0 == pid || 4 == pid)
         return psutil_boot_time(NULL, NULL);
-    }
 
     hProcess = psutil_handle_from_pid(pid);
-    if (hProcess == NULL) {
+    if (hProcess == NULL)
         return NULL;
-    }
-
     if (! GetProcessTimes(hProcess, &ftCreate, &ftExit, &ftKernel, &ftUser)) {
         CloseHandle(hProcess);
         if (GetLastError() == ERROR_ACCESS_DENIED) {
@@ -513,9 +472,8 @@ psutil_proc_create_time(PyObject *self, PyObject *args)
     ret = GetExitCodeProcess(hProcess, &exitCode);
     CloseHandle(hProcess);
     if (ret != 0) {
-        if (exitCode != STILL_ACTIVE) {
+        if (exitCode != STILL_ACTIVE)
             return NoSuchProcess();
-        }
     }
     else {
         // Ignore access denied as it means the process is still alive.
@@ -538,39 +496,6 @@ psutil_proc_create_time(PyObject *self, PyObject *args)
 }
 
 
-/*
- * Alternative implementation of the one above but bypasses ACCESS DENIED.
- */
-static PyObject *
-psutil_proc_create_time_2(PyObject *self, PyObject *args)
-{
-    DWORD pid;
-    PSYSTEM_PROCESS_INFORMATION process;
-    PVOID buffer;
-    long long   unix_time;
-
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
-        return NULL;
-    }
-    if (! psutil_get_proc_info(pid, &process, &buffer)) {
-        return NULL;
-    }
-    // special case for PIDs 0 and 4, return system boot time
-    if (0 == pid || 4 == pid) {
-        return psutil_boot_time(NULL, NULL);
-    }
-    /*
-    Convert the LARGE_INTEGER union to a Unix time.
-    It's the best I could find by googling and borrowing code here and there.
-    The time returned has a precision of 1 second.
-    */
-    unix_time = ((LONGLONG)process->CreateTime.HighPart) << 32;
-    unix_time += process->CreateTime.LowPart - 116444736000000000LL;
-    unix_time /= 10000000;
-    free(buffer);
-    return Py_BuildValue("d", (double)unix_time);
-}
-
 
 /*
  * Return the number of logical CPUs.
@@ -582,13 +507,10 @@ psutil_cpu_count_logical(PyObject *self, PyObject *args)
     system_info.dwNumberOfProcessors = 0;
 
     GetSystemInfo(&system_info);
-    if (system_info.dwNumberOfProcessors == 0) {
-        // mimic os.cpu_count()
-        Py_RETURN_NONE;
-    }
-    else {
+    if (system_info.dwNumberOfProcessors == 0)
+        Py_RETURN_NONE;  // mimic os.cpu_count()
+    else
         return Py_BuildValue("I", system_info.dwNumberOfProcessors);
-    }
 }
 
 
@@ -664,20 +586,16 @@ psutil_proc_cmdline(PyObject *self, PyObject *args) {
     int pid_return;
     PyObject *arglist;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
-    if ((pid == 0) || (pid == 4)) {
+    if ((pid == 0) || (pid == 4))
         return Py_BuildValue("[]");
-    }
 
     pid_return = psutil_pid_is_running(pid);
-    if (pid_return == 0) {
+    if (pid_return == 0)
         return NoSuchProcess();
-    }
-    if (pid_return == -1) {
+    if (pid_return == -1)
         return NULL;
-    }
 
     // XXX the assumptio below probably needs to go away
 
@@ -703,14 +621,12 @@ psutil_proc_exe(PyObject *self, PyObject *args) {
     long pid;
     HANDLE hProcess;
     wchar_t exe[MAX_PATH];
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
 
     hProcess = psutil_handle_from_pid_waccess(pid, PROCESS_QUERY_INFORMATION);
-    if (NULL == hProcess) {
+    if (NULL == hProcess)
         return NULL;
-    }
     if (GetProcessImageFileNameW(hProcess, exe, MAX_PATH) == 0) {
         CloseHandle(hProcess);
         PyErr_SetFromWindowsErr(0);
@@ -718,6 +634,47 @@ psutil_proc_exe(PyObject *self, PyObject *args) {
     }
     CloseHandle(hProcess);
     return Py_BuildValue("u", exe);
+}
+
+
+/*
+ * Return process base name.
+ * Note: psutil_proc_exe() is attempted first because it's faster
+ * but it raise AccessDenied for processes owned by other users
+ * in which case we fall back on using this.
+ */
+static PyObject *
+psutil_proc_name(PyObject *self, PyObject *args) {
+    long pid;
+    int ok;
+    PROCESSENTRY32 pentry;
+    HANDLE hSnapShot;
+
+    if (! PyArg_ParseTuple(args, "l", &pid))
+        return NULL;
+    hSnapShot = CreateToolhelp32Snapshot(TH32CS_SNAPPROCESS, pid);
+    if (hSnapShot == INVALID_HANDLE_VALUE) {
+        PyErr_SetFromWindowsErr(0);
+        return NULL;
+    }
+    pentry.dwSize = sizeof(PROCESSENTRY32);
+    ok = Process32First(hSnapShot, &pentry);
+    if (! ok) {
+        CloseHandle(hSnapShot);
+        PyErr_SetFromWindowsErr(0);
+        return NULL;
+    }
+    while (ok) {
+        if (pentry.th32ProcessID == pid) {
+            CloseHandle(hSnapShot);
+            return Py_BuildValue("s", pentry.szExeFile);
+        }
+        ok = Process32Next(hSnapShot, &pentry);
+    }
+
+    CloseHandle(hSnapShot);
+    NoSuchProcess();
+    return NULL;
 }
 
 
@@ -736,18 +693,15 @@ psutil_proc_memory_info(PyObject *self, PyObject *args)
 #endif
     SIZE_T private = 0;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
 
     hProcess = psutil_handle_from_pid(pid);
-    if (NULL == hProcess) {
+    if (NULL == hProcess)
         return NULL;
-    }
 
     if (! GetProcessMemoryInfo(hProcess, (PPROCESS_MEMORY_COUNTERS)&cnt,
-                               sizeof(cnt)) )
-    {
+                               sizeof(cnt))) {
         CloseHandle(hProcess);
         return PyErr_SetFromWindowsErr(0);
     }
@@ -810,12 +764,10 @@ psutil_proc_memory_info_2(PyObject *self, PyObject *args)
     unsigned int m1, m2, m3, m4, m5, m6, m7, m8;
 #endif
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
-    if (! psutil_get_proc_info(pid, &process, &buffer)) {
+    if (! psutil_get_proc_info(pid, &process, &buffer))
         return NULL;
-    }
 
 #if (_WIN32_WINNT >= 0x0501)  // Windows XP with SP2
     private = process->PrivatePageCount;
@@ -858,10 +810,8 @@ psutil_virtual_mem(PyObject *self, PyObject *args)
     MEMORYSTATUSEX memInfo;
     memInfo.dwLength = sizeof(MEMORYSTATUSEX);
 
-    if (! GlobalMemoryStatusEx(&memInfo) ) {
+    if (! GlobalMemoryStatusEx(&memInfo))
         return PyErr_SetFromWindowsErr(0);
-    }
-
     return Py_BuildValue("(LLLLLL)",
                          memInfo.ullTotalPhys,      // total
                          memInfo.ullAvailPhys,      // avail
@@ -883,9 +833,8 @@ psutil_cpu_times(PyObject *self, PyObject *args)
     float idle, kernel, user, system;
     FILETIME idle_time, kernel_time, user_time;
 
-    if (!GetSystemTimes(&idle_time, &kernel_time, &user_time)) {
+    if (!GetSystemTimes(&idle_time, &kernel_time, &user_time))
         return PyErr_SetFromWindowsErr(0);
-    }
 
     idle = (float)((HI_T * idle_time.dwHighDateTime) + \
                    (LO_T * idle_time.dwLowDateTime));
@@ -986,12 +935,10 @@ psutil_per_cpu_times(PyObject *self, PyObject *args)
 error:
     Py_XDECREF(arg);
     Py_DECREF(retlist);
-    if (sppi) {
+    if (sppi)
         free(sppi);
-    }
-    if (hNtDll) {
+    if (hNtDll)
         FreeLibrary(hNtDll);
-    }
     PyErr_SetFromWindowsErr(0);
     return NULL;
 }
@@ -1014,14 +961,12 @@ psutil_proc_cwd(PyObject *self, PyObject *args)
     PyObject *cwd_from_wchar = NULL;
     PyObject *cwd = NULL;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
 
     processHandle = psutil_handle_from_pid(pid);
-    if (processHandle == NULL) {
+    if (processHandle == NULL)
         return NULL;
-    }
 
     pebAddress = psutil_get_peb_address(processHandle);
 
@@ -1210,13 +1155,11 @@ psutil_proc_suspend(PyObject *self, PyObject *args)
 {
     long pid;
     int suspend = 1;
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
-        return NULL;
-    }
 
-    if (! psutil_proc_suspend_or_resume(pid, suspend)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
+    if (! psutil_proc_suspend_or_resume(pid, suspend))
+        return NULL;
     Py_RETURN_NONE;
 }
 
@@ -1226,34 +1169,12 @@ psutil_proc_resume(PyObject *self, PyObject *args)
 {
     long pid;
     int suspend = 0;
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
-        return NULL;
-    }
 
-    if (! psutil_proc_suspend_or_resume(pid, suspend)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
+    if (! psutil_proc_suspend_or_resume(pid, suspend))
+        return NULL;
     Py_RETURN_NONE;
-}
-
-
-static PyObject *
-psutil_proc_num_threads(PyObject *self, PyObject *args)
-{
-    DWORD pid;
-    PSYSTEM_PROCESS_INFORMATION process;
-    PVOID buffer;
-    int num;
-
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
-        return NULL;
-    }
-    if (! psutil_get_proc_info(pid, &process, &buffer)) {
-        return NULL;
-    }
-    num = (int)process->NumberOfThreads;
-    free(buffer);
-    return Py_BuildValue("i", num);
 }
 
 
@@ -1270,12 +1191,10 @@ psutil_proc_threads(PyObject *self, PyObject *args)
     PyObject *pyTuple = NULL;
     HANDLE hThreadSnap = NULL;
 
-    if (retList == NULL) {
+    if (retList == NULL)
         return NULL;
-    }
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         goto error;
-    }
     if (pid == 0) {
         // raise AD instead of returning 0 as procexp is able to
         // retrieve useful information somehow
@@ -1288,9 +1207,8 @@ psutil_proc_threads(PyObject *self, PyObject *args)
         NoSuchProcess();
         goto error;
     }
-    if (pid_return == -1) {
+    if (pid_return == -1)
         goto error;
-    }
 
     hThreadSnap = CreateToolhelp32Snapshot(TH32CS_SNAPTHREAD, 0);
     if (hThreadSnap == INVALID_HANDLE_VALUE) {
@@ -1308,10 +1226,8 @@ psutil_proc_threads(PyObject *self, PyObject *args)
 
     // Walk the thread snapshot to find all threads of the process.
     // If the thread belongs to the process, increase the counter.
-    do
-    {
-        if (te32.th32OwnerProcessID == pid)
-        {
+    do {
+        if (te32.th32OwnerProcessID == pid) {
             pyTuple = NULL;
             hThread = NULL;
             hThread = OpenThread(THREAD_QUERY_INFORMATION,
@@ -1362,9 +1278,8 @@ error:
     Py_DECREF(retList);
     if (hThread != NULL)
         CloseHandle(hThread);
-    if (hThreadSnap != NULL) {
+    if (hThreadSnap != NULL)
         CloseHandle(hThreadSnap);
-    }
     return NULL;
 }
 
@@ -1377,20 +1292,16 @@ psutil_proc_open_files(PyObject *self, PyObject *args)
     DWORD      access = PROCESS_DUP_HANDLE | PROCESS_QUERY_INFORMATION;
     PyObject  *filesList;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
 
     processHandle = psutil_handle_from_pid_waccess(pid, access);
-    if (processHandle == NULL) {
+    if (processHandle == NULL)
         return NULL;
-    }
-
     filesList = psutil_get_open_files(pid, processHandle);
     CloseHandle(processHandle);
-    if (filesList == NULL) {
+    if (filesList == NULL)
         return PyErr_SetFromWindowsErr(0);
-    }
     return filesList;
 }
 
@@ -1407,12 +1318,10 @@ psutil_win32_QueryDosDevice(PyObject *self, PyObject *args)
     TCHAR d = TEXT('A');
     TCHAR     szBuff[5];
 
-    if (!PyArg_ParseTuple(args, "s", &lpDevicePath)) {
+    if (!PyArg_ParseTuple(args, "s", &lpDevicePath))
         return NULL;
-    }
 
-    while (d <= TEXT('Z'))
-    {
+    while (d <= TEXT('Z')) {
         TCHAR szDeviceName[3] = {d, TEXT(':'), TEXT('\0')};
         TCHAR szTarget[512] = {0};
         if (QueryDosDevice(szDeviceName, szTarget, 511) != 0) {
@@ -1446,15 +1355,13 @@ psutil_proc_username(PyObject *self, PyObject *args)
     PTSTR fullName;
     PyObject *returnObject;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
 
     processHandle = psutil_handle_from_pid_waccess(
         pid, PROCESS_QUERY_INFORMATION);
-    if (processHandle == NULL) {
+    if (processHandle == NULL)
         return NULL;
-    }
 
     if (!OpenProcessToken(processHandle, TOKEN_QUERY, &tokenHandle)) {
         CloseHandle(processHandle);
@@ -1467,9 +1374,8 @@ psutil_proc_username(PyObject *self, PyObject *args)
 
     bufferSize = 0x100;
     user = malloc(bufferSize);
-    if (user == NULL) {
+    if (user == NULL)
         return PyErr_NoMemory();
-    }
 
     if (!GetTokenInformation(tokenHandle, TokenUser, user, bufferSize,
                              &bufferSize))
@@ -2001,9 +1907,8 @@ psutil_proc_priority_get(PyObject *self, PyObject *args)
     long pid;
     DWORD priority;
     HANDLE hProcess;
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
 
     hProcess = psutil_handle_from_pid(pid);
     if (hProcess == NULL) {
@@ -2066,9 +1971,8 @@ psutil_proc_io_priority_get(PyObject *self, PyObject *args)
         (_NtQueryInformationProcess)GetProcAddress(
             GetModuleHandleA("ntdll.dll"), "NtQueryInformationProcess");
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
     hProcess = psutil_handle_from_pid(pid);
     if (hProcess == NULL) {
         return NULL;
@@ -2137,9 +2041,8 @@ psutil_proc_io_counters(PyObject *self, PyObject *args)
     HANDLE hProcess;
     IO_COUNTERS IoCounters;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
     hProcess = psutil_handle_from_pid(pid);
     if (NULL == hProcess) {
         return NULL;
@@ -2158,32 +2061,6 @@ psutil_proc_io_counters(PyObject *self, PyObject *args)
 
 
 /*
- * Alternative implementation of the one above but bypasses ACCESS DENIED.
- */
-static PyObject *
-psutil_proc_io_counters_2(PyObject *self, PyObject *args)
-{
-    DWORD pid;
-    PSYSTEM_PROCESS_INFORMATION process;
-    PVOID buffer;
-    LONGLONG rcount, wcount, rbytes, wbytes;
-
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
-        return NULL;
-    }
-    if (! psutil_get_proc_info(pid, &process, &buffer)) {
-        return NULL;
-    }
-    rcount = process->ReadOperationCount.QuadPart;
-    wcount = process->WriteOperationCount.QuadPart;
-    rbytes = process->ReadTransferCount.QuadPart;
-    wbytes = process->WriteTransferCount.QuadPart;
-    free(buffer);
-    return Py_BuildValue("KKKK", rcount, wcount, rbytes, wbytes);
-}
-
-
-/*
  * Return process CPU affinity as a bitmask
  */
 static PyObject *
@@ -2194,9 +2071,8 @@ psutil_proc_cpu_affinity_get(PyObject *self, PyObject *args)
     DWORD_PTR proc_mask;
     DWORD_PTR system_mask;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
     hProcess = psutil_handle_from_pid(pid);
     if (hProcess == NULL) {
         return NULL;
@@ -2262,9 +2138,8 @@ psutil_proc_is_suspended(PyObject *self, PyObject *args)
     PSYSTEM_PROCESS_INFORMATION process;
     PVOID buffer;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
     if (! psutil_get_proc_info(pid, &process, &buffer)) {
         return NULL;
     }
@@ -2793,9 +2668,8 @@ psutil_proc_num_handles(PyObject *self, PyObject *args)
     HANDLE hProcess;
     DWORD handleCount;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
     hProcess = psutil_handle_from_pid(pid);
     if (NULL == hProcess) {
         return NULL;
@@ -2809,52 +2683,66 @@ psutil_proc_num_handles(PyObject *self, PyObject *args)
 }
 
 
-/*
- * Alternative implementation of the one above but bypasses ACCESS DENIED.
- */
 static PyObject *
-psutil_proc_num_handles_2(PyObject *self, PyObject *args)
+psutil_proc_info(PyObject *self, PyObject *args)
 {
     DWORD pid;
     PSYSTEM_PROCESS_INFORMATION process;
     PVOID buffer;
-    ULONG count;
-
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
-        return NULL;
-    }
-    if (! psutil_get_proc_info(pid, &process, &buffer)) {
-        return NULL;
-    }
-    count = process->HandleCount;
-    free(buffer);
-    return Py_BuildValue("k", count);
-}
-
-
-/*
- * Return the number of context switches executed by process.
- */
-static PyObject *
-psutil_proc_num_ctx_switches(PyObject *self, PyObject *args)
-{
-    DWORD pid;
-    PSYSTEM_PROCESS_INFORMATION process;
-    PVOID buffer;
+    ULONG num_handles;
     ULONG i;
-    ULONG total = 0;
+    ULONG ctx_switches = 0;
+    double user_time;
+    double kernel_time;
+    long long create_time;
+    int num_threads;
+    LONGLONG io_rcount, io_wcount, io_rbytes, io_wbytes;
 
-    if (! PyArg_ParseTuple(args, "l", &pid)) {
+
+    if (! PyArg_ParseTuple(args, "l", &pid))
         return NULL;
-    }
-    if (! psutil_get_proc_info(pid, &process, &buffer)) {
+    if (! psutil_get_proc_info(pid, &process, &buffer))
         return NULL;
+
+    num_handles = process->HandleCount;
+    for (i = 0; i < process->NumberOfThreads; i++)
+        ctx_switches += process->Threads[i].ContextSwitches;
+    user_time = (double)process->UserTime.HighPart * 429.4967296 + \
+                (double)process->UserTime.LowPart * 1e-7;
+    kernel_time = (double)process->KernelTime.HighPart * 429.4967296 + \
+                  (double)process->KernelTime.LowPart * 1e-7;
+    // Convert the LARGE_INTEGER union to a Unix time.
+    // It's the best I could find by googling and borrowing code here
+    // and there. The time returned has a precision of 1 second.
+    if (0 == pid || 4 == pid) {
+        // the python module will translate this into BOOT_TIME later
+        create_time = 0;
     }
-    for (i = 0; i < process->NumberOfThreads; i++) {
-        total += process->Threads[i].ContextSwitches;
+    else {
+        create_time = ((LONGLONG)process->CreateTime.HighPart) << 32;
+        create_time += process->CreateTime.LowPart - 116444736000000000LL;
+        create_time /= 10000000;
     }
+    num_threads = (int)process->NumberOfThreads;
+    io_rcount = process->ReadOperationCount.QuadPart;
+    io_wcount = process->WriteOperationCount.QuadPart;
+    io_rbytes = process->ReadTransferCount.QuadPart;
+    io_wbytes = process->WriteTransferCount.QuadPart;
     free(buffer);
-    return Py_BuildValue("ki", total, 0);
+
+    return Py_BuildValue(
+        "kkdddiKKKK",
+        num_handles,
+        ctx_switches,
+        user_time,
+        kernel_time,
+        (double)create_time,
+        num_threads,
+        io_rcount,
+        io_wcount,
+        io_rbytes,
+        io_wbytes
+    );
 }
 
 
@@ -3281,6 +3169,8 @@ PsutilMethods[] =
      "Return process cmdline as a list of cmdline arguments"},
     {"proc_exe", psutil_proc_exe, METH_VARARGS,
      "Return path of the process executable"},
+    {"proc_name", psutil_proc_name, METH_VARARGS,
+     "Return process name"},
     {"proc_kill", psutil_proc_kill, METH_VARARGS,
      "Kill the process identified by the given PID"},
     {"proc_cpu_times", psutil_proc_cpu_times, METH_VARARGS,
@@ -3290,6 +3180,8 @@ PsutilMethods[] =
      "seconds since the epoch"},
     {"proc_memory_info", psutil_proc_memory_info, METH_VARARGS,
      "Return a tuple of process memory information"},
+    {"proc_memory_info_2", psutil_proc_memory_info, METH_VARARGS,
+     "Alternate implementation"},
     {"proc_cwd", psutil_proc_cwd, METH_VARARGS,
      "Return process current working directory"},
     {"proc_suspend", psutil_proc_suspend, METH_VARARGS,
@@ -3300,8 +3192,6 @@ PsutilMethods[] =
      "Return files opened by process"},
     {"proc_username", psutil_proc_username, METH_VARARGS,
      "Return the username of a process"},
-    {"proc_num_threads", psutil_proc_num_threads, METH_VARARGS,
-     "Return the network connections of a process"},
     {"proc_threads", psutil_proc_threads, METH_VARARGS,
      "Return process threads information as a list of tuple"},
     {"proc_wait", psutil_proc_wait, METH_VARARGS,
@@ -3326,22 +3216,12 @@ PsutilMethods[] =
      "Return True if one of the process threads is in a suspended state"},
     {"proc_num_handles", psutil_proc_num_handles, METH_VARARGS,
      "Return the number of handles opened by process."},
-    {"proc_num_ctx_switches", psutil_proc_num_ctx_switches, METH_VARARGS,
-     "Return the number of context switches performed by process."},
     {"proc_memory_maps", psutil_proc_memory_maps, METH_VARARGS,
      "Return a list of process's memory mappings"},
 
     // --- alternative pinfo interface
-    {"proc_cpu_times_2", psutil_proc_cpu_times_2, METH_VARARGS,
-     "Alternative implementation"},
-    {"proc_create_time_2", psutil_proc_create_time_2, METH_VARARGS,
-     "Alternative implementation"},
-    {"proc_num_handles_2", psutil_proc_num_handles_2, METH_VARARGS,
-     "Alternative implementation"},
-    {"proc_io_counters_2", psutil_proc_io_counters_2, METH_VARARGS,
-     "Alternative implementation"},
-    {"proc_memory_info_2", psutil_proc_memory_info_2, METH_VARARGS,
-     "Alternative implementation"},
+    {"proc_info", psutil_proc_info, METH_VARARGS,
+     "Various process information"},
 
     // --- system-related functions
     {"pids", psutil_pids, METH_VARARGS,